pyproject.toml
[project]
name = "mini-docs-mcp"
version = "0.1.0"
description = "MCP server for intelligent Markdown knowledge base with semantic search and bidirectional linking"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Mini Docs MCP Contributors"}
]
keywords = ["mcp", "markdown", "rag", "semantic-search", "knowledge-base", "backlinks"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastmcp>=2.0.0",
    "lancedb>=0.15.0",
    "sentence-transformers>=3.0.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0",
    "watchdog>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.8.0",
]

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/mini-docs-mcp"
Documentation = "https://github.com/YOUR_USERNAME/mini-docs-mcp#readme"
Repository = "https://github.com/YOUR_USERNAME/mini-docs-mcp"
Issues = "https://github.com/YOUR_USERNAME/mini-docs-mcp/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mini_docs_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["mini_docs_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra -q"

[tool.coverage.run]
source = ["src/mini_docs_mcp"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[project.scripts]
mini-docs-mcp = "mini_docs_mcp.server:main"
